backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
flask-jwt-extended==4.6.0
APScheduler==3.10.4
Flask-Limiter==3.5.0
groq>=0.4.0
